add separate directory component
