+++ plugins/Files/js/components/directory.js
@@ -1,0 +1,17 @@
+import React, { PropTypes } from 'react'
+
+const Directory = ({onClick, name}) => (
+	<li onClick={onClick}>
+		<div className="filename">
+			<i className="fa fa-folder"></i>
+			{name}
+		</div>
+	</li>
+)
+
+Directory.propTypes = {
+	onClick: PropTypes.func.isRequired,
+	name: PropTypes.string.isRequired,
+}
+
+export default Directory
