+++ plugins/Files/js/containers/filelist.js
@@ -1,7 +1,7 @@
 import FileListView from '../components/filelist.js'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
-import { setPath } from '../actions/files.js'
+import { setPath, downloadFile, deleteFile } from '../actions/files.js'
 
 const mapStateToProps = (state) => ({
 	files: state.files.get('workingDirectoryFiles'),
@@ -10,7 +10,7 @@ const mapStateToProps = (state) => ({
 	showSearchField: state.files.get('showSearchField'),
 })
 const mapDispatchToProps = (dispatch) => ({
-	actions: bindActionCreators({ setPath }, dispatch),
+	actions: bindActionCreators({ setPath, downloadFile, deleteFile }, dispatch),
 })
 
 const FileList = connect(mapStateToProps, mapDispatchToProps)(FileListView)
